+++ src/git/exports.ts
@@ -1,0 +1,1 @@
+export { AppendNote, AppendNoteParameters } from "./append";